Market model simplification     -now takes Constants as market key     -hidden data requests
Exchange takes multiple models and is provided to agents.
It plays a role of global market where agents can work with multiple
market models

+++ src/python/marketsim/MarketModel.py
@@ -1,66 +1,77 @@
 # MarketModel module
 import calendar
-import requests, os
+import requests, os, sys
 import json
 from dateutil.parser import parse
 from datetime import datetime, timedelta, date
 import matplotlib.pyplot as plt
 import quandl, tokens
 import pandas as pd
+from marketsim.Constants import CCYMARKET, CCY
+from pathlib import Path
 
 def daterange(start_date, end_date):
-    for n in range(int ((end_date - start_date).days)):
+    for n in range(int((end_date - start_date).days)):
         yield start_date + timedelta(n)
 
 """
 MarketModel base class for modelling a FX market
 """
 class MarketModel:
-    url = "http://api.fixer.io/" # Fixer was replaced with Quandl info source, due to limitations of Fixer
     data_path = "data/"
 
-    def __init__(self):
-        self.agency='ECB'
-        self.baseCurrency = "USD"
-        self.quoteValue = "EUR"  # quote value is not only a currency but also a price e.g. of commodity share
-        self.fileName = self.baseCurrency+'-'+self.quoteValue
+    def __init__(self, requested_market=None):
         self.startDate = parse("2000-01-01")
         self.endDate = datetime.today()
         self.marketprice = {}
+        self.market_key = None
+        self.df = None
+        if requested_market is not None:
+            self.set_market(requested_market)
+
+    def set_market(self, requested_market):
+        self.market_key = requested_market
+        if type(self.market_key) == CCYMARKET:
+            self.get_quandl_data()
+
+    def __str__(self):
+        return self.base_currency_name() + '-' + self.quote_value_name()
+
+    def base_currency_name(self):  #e.g. USD/EUR/etc
+        if type(self.market_key) == CCYMARKET:
+            if type(self.market_key) == CCYMARKET:
+                return self.market_key.value[0].name
+        else:
+            print("Unexpected market data request")
+            raise
+
+    def quote_value_name(self):  #e.g. GOLD/OIL/GOOGLE/etc
+        if type(self.market_key) == CCYMARKET:
+            if type(self.market_key) == CCYMARKET:
+                return self.market_key.value[1].name
+        else:
+            print("Unexpected market data request")
+            raise
+
+    def get_quandl_data(self):
+        agency = 'ECB'
+        file_name = self.data_path + self.__str__() + ".csv"
+        data_file = Path(file_name)
+        if data_file.is_file():
+            self.df = pd.read_csv(file_name)
+            print("read from file: " + file_name)
+        else:
+            self.df = quandl.get(agency+'/'+self.base_currency_name()+self.quote_value_name(), authtoken=tokens.get_quandl_tocken())
+            if not os.path.exists(self.data_path):
+                os.makedirs(self.data_path)
+            self.df.to_csv(file_name, index=True, sep='\t')
+            print("saved to file: " + file_name)
 
     def start_date(self, date_str):
         self.startDate = parse(date_str)
 
     def end_date(self, date_str):
         self.endDate = parse(date_str)
-
-    def request(self):
-        self.df = quandl.get(self.agency+'/'+self.quoteValue+self.baseCurrency, authtoken=tokens.get_quandl_tocken())
-        if not os.path.exists(self.data_path):
-            os.makedirs(self.data_path)
-        file_name = self.data_path + self.fileName + ".csv"
-        self.df.to_csv(file_name, index=True, sep='\t')
-        print("saved to file: " + file_name)
-
-    def requestFixer(self):
-        # date_range = daterange(self.startDate, self.endDate) # does not work with fixer- request rate is limited
-        date_range = daterange(self.endDate - timedelta(10), self.endDate)  # TODO change - avoid fixer.io
-        if not os.path.exists(self.data_path):
-            os.makedirs(self.data_path)
-        f = open(self.data_path + self.fileName + ".csv", 'w')
-        f.write("DATE\t" + self.fileName + "\n")
-        for d in date_range:
-            date_str = '{:%Y-%m-%d}'.format(d)
-            req_str = self.url + date_str + "?base=" + self.baseCurrency
-            try:
-                r = requests.get(req_str)
-                f.write(date_str + "\t{}\n".format(r.json()['rates'][self.quoteValue]))
-                self.marketprice['{:%Y-%m-%d}'.format(d)] = r.json()['rates'][self.quoteValue]
-            except:
-                print('Problem with date: ' + req_str)
-                raise
-        f.close()
-        print("saved to file: " + self.fileName)
 
     def get_market_price(self, timestamp):
         try:
@@ -93,6 +104,6 @@ class PerfectFXModel(MarketModel):
 if __name__ == "__main__":
     print("Market Model test")
     mm = MarketModel()
-    mm.request()
+    mm.collect_market_data()
     rate_this_day = mm.get_market_price('2017-07-19')
     print(rate_this_day)
